Renamed and refactored(without affecting on asserts) test
